// apps/portal-api/src/model/kanel/public/Competitor.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

/** Identifier type for public.Competitor */
export type CompetitorId = string & {
  __brand: 'public.Competitor';
};

/** Represents the table public.Competitor */
export default interface Competitor {
  id: string | null;

  name: string | null;

  tier: string | null;

  domain: string | null;
}

/** Represents the initializer for the table public.Competitor */
export interface CompetitorInitializer {
  /** Default value: gen_random_uuid() */
  id?: string | null;

  name?: string | null;

  tier?: string | null;

  domain?: string | null;
}

/** Represents the mutator for the table public.Competitor */
export interface CompetitorMutator {
  id?: string | null;

  name?: string | null;

  tier?: string | null;

  domain?: string | null;
}